# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Article API",
    "description": "Article API to manage Admin users, Tenants, Articles, Tags and References.",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/api/",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/tenants": {
      "post": {
        "tags": ["Tenant"],
        "description": "Create new Tenant",
        "operationId": "createTenant",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Tenant Parameters",
            "in": "body",
            "description": "Create new Tenant",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "403": {
            "description": "Tenant Login Duplication",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },      
      "get": {
        "tags": ["Tenant"],
        "description": "Get Tenant List",
        "operationId": "getTenants",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Get Tenant List",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Tenant List",
            "schema": {
              "$ref": "#/definitions/GetUser"
            }
          },
          "403": {
            "description": "Tenant Login Duplication",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/tenants/{login}": {
      "patch": {
        "tags": ["Tenant"],
        "description": "Update Tenant",
        "operationId": "updateTenant",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Tenant Parameters",
            "in": "body",
            "description": "Update Tenant",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          },
          {
            "name": "login",
            "in": "path",
            "description": "Login",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "405": {
            "description": "Missing Parameter",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "Invalid login",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "delete": {
        "tags": ["Tenant"],
        "description": "Delete Tenant",
        "operationId": "deleteTenant",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "login",
            "in": "path",
            "description": "Login",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Success message",
            "schema": {
              "type": "Response with Success message"
            }
          },
          "404": {
            "description": "Invalid login",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/tenants/accesskey": {
      "get": {
        "tags": ["Tenant"],
        "description": "Get Access Key of Tenant",
        "operationId": "getAccessKey",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Response with Access Key",
            "schema": {
              "$ref": "#/definitions/TenantAccessKey"
            }
          },
          "404": {
            "description": "Invalid login",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Authentication Failed",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdateUser": {
      "properties": {
        "name": {
          "required":true,
          "type": "string",
          "example": "Erik Clasie"
        },
        "token": {
          "required":true,
          "type": "string",
          "example": "Token"
        }
      }
    },
    "GetUser": {
      "properties": {
        "name": {
          "required":true,
          "type": "string",
          "example": "Erik Clasie"
        },
        "login": {
          "required":true,
          "type": "string",
          "example": "login"
        }
      }
    },
...
    "User": {
      "properties": {
        "name": {
          "required":true,
          "type": "string",
          "example": "Erik Clasie"
        },
        "login": {
          "required":true,
          "type": "string",
          "example": "erik"
        },
        "password": {
          "required":true,
          "type":"string",
          "example": "password"
        },
        "token": {
          "required":true,
          "type":"string",
          "example": "Token"
        } 
      }
    },
    "TenantAccessKey": {
      "properties": {
        "access_key": {
          "required":true,
          "type": "string",
          "example": "Access Key"
        }
      }
    },
...
  }
}